lambda_schedule_roi/python/requirements.txt
Flask==2.3.2
numpy==1.24.3
pandas==2.0.1
yfinance==0.2.18
aws_wsgi==0.2.7
boto3==1.28.62
Flask_Cors==4.0.0
python-dotenv==1.0.0
requests-html==0.10.0
python-math==0.0.1
schedule==1.2.1
